Rename to get rid of right-arrow bar
Totally a typo from me, when kind of using the arrow keys on the
keyboard..

From A GPT

The sequence \342\200\223 in the filename you've provided is an example
of an octal escape sequence. In many programming languages and systems,
octal escape sequences are used to represent characters that might not
be easily typed on a keyboard or to represent characters that have
special meanings in the context of the language or system.

The octal escape sequence \342\200\223 represents a Unicode character.
To understand what character it represents, we need to convert the octal
values to decimal and then look up the corresponding Unicode character.

Here's how you can do it:

Convert each octal value to decimal:
\342 in octal is 226 in decimal.
\200 in octal is 128 in decimal.
\223 in octal is 157 in decimal.
Combine these decimal values to get the Unicode code point: 226128157.
Look up the Unicode character with the code point 226128157.
Upon looking up the Unicode character with the code point 226128157, it
corresponds to the character "‣" (a rightwards arrow to bar). This
character is often used in text to indicate a continuation or a
transition to a new section or part.

So, in the context of your filename, \342\200\223 represents the "‣"
character, which is likely used to indicate a continuation or a
transition to a new part of the content, similar to how a hyphen might
be used in a title or filename to indicate a continuation or a sequel.